+++ .eslintrc.js
@@ -1,4 +1,7 @@
 module.exports = {
   root: true,
   extends: '@react-native',
+  rules: {
+    'react/react-in-jsx-scope': 'off',
+  },
 };
