+++ constants/index.js
@@ -1,0 +1,1 @@
+export const SOURCE_URL = 'https://gloddy.vercel.app';
